+++ Configuration/TCA/Overrides/tx_cfcleague_competition.php
@@ -22,6 +22,31 @@ $columns = [
             ],
         ],
     ],
+    'statsenabled' => [
+        'exclude' => 1,
+        'label' => 'LLL:EXT:t3sportstats/Resources/Private/Language/locallang_db.xlf:tx_t3sportstats_stats_enabled',
+        'config' => [
+            'type' => 'check',
+            'default' => '0',
+        ],
+    ],
+    'statsrefresh' => [
+        'exclude' => 1,
+        'label' => 'LLL:EXT:t3sportstats/Resources/Private/Language/locallang_db.xlf:tx_t3sportstats_stats_refresh',
+        'config' => [
+            'type' => 'check',
+            'default' => '0',
+        ],
+    ],
+    'statsrefreshed' => [
+        'exclude' => 1,
+        'label' => 'LLL:EXT:t3sportstats/Resources/Private/Language/locallang_db.xlf:tx_t3sportstats_stats_refreshed',
+        'config' => [
+            'type' => 'input',
+            'size' => '20',
+            'readOnly' => '1',
+        ],
+    ],
 ];
 
 $table = [
@@ -35,4 +60,4 @@ Sys25\RnBase\Backend\Utility\TcaTool::configureWizards($table, [
 ]);
 
 Sys25\RnBase\Utility\Extensions::addTCAcolumns('tx_cfcleague_competition', $columns, 1);
-Sys25\RnBase\Utility\Extensions::addToAllTCAtypes('tx_cfcleague_competition', 'tags', '', 'after:point_system');
+Sys25\RnBase\Utility\Extensions::addToAllTCAtypes('tx_cfcleague_competition', 'tags, statsenabled, statsrefresh, statsrefreshed', '', 'after:point_system');